+++ vm/src/tests/cairo_pie_comparator.py
@@ -23,5 +23,3 @@ with ZipFile(filename1) as cairo_lang_pie_zip, ZipFile(filename2) as cairo_vm_pi
     # Compare binary files
     with cairo_lang_pie_zip.open("memory.bin", 'r') as f1,  cairo_vm_pie_zip.open("memory.bin", 'r') as f2:
         memory_comparator.compare_memory_file_contents(f1.read(), f2.read())
-
-    print(f"Comparison succesful for {filename1} vs {filename2}")